+++ R/useShinyalert.R
@@ -1,12 +1,17 @@
 #' Set up a Shiny app to use shinyalert
 #'
-#' This function must be called from a Shiny app's UI in order for the
-#' \code{\link[shinyalert]{shinyalert}} function to work.\cr\cr
-#' You can call \code{useShinyalert()} from anywhere inside the UI.
+#' **This function is no longer needed.**\cr\cr
+#' When a {shinyalert} message is shown for the first time,
+#' the required scripts are automatically inserted to the Shiny app. If you want to pre-load
+#' these scripts into the app for performance reasons, you can call this function anywhere
+#' in the UI and use `force = TRUE`.\cr\cr
+#' Note that it is no longer needed to use this function.
 #'
 #' @param rmd Set this to \code{TRUE} if using \code{shinyalert}
 #' inside an interactive R markdown document. The YAML of the Rmd file must have
 #' \code{runtime: shiny}.
+#' @param force Set to `TRUE` to load the {shinyalert} scripts. If `FALSE` (default),
+#' you will get a warning saying this function is not required.
 #' @return Scripts that \code{shinyalert} requires that are automatically
 #' inserted to the app's \code{<head>} tag.
 #' @examples
@@ -16,7 +21,7 @@
 #'
 #'   shinyApp(
 #'     ui = fluidPage(
-#'       useShinyalert(),  # Set up shinyalert
+#'       useShinyalert(force = TRUE),  # Set up shinyalert
 #'       actionButton("btn", "Click me")
 #'     ),
 #'     server = function(input, output) {
@@ -29,35 +34,29 @@
 #' }
 #' @seealso \code{\link[shinyalert]{shinyalert}}
 #' @export
-useShinyalert <- function(rmd = FALSE) {
-  stopifnot(rmd == TRUE || rmd == FALSE)
-
-  shiny::addResourcePath("resources",
-                         system.file("www", package = "shinyalert"))
-  insert_into_doc <- if (rmd) shiny::tagList else shiny::tags$head
-
-  shiny::singleton(
-    insert_into_doc(
-      shiny::tags$script(
-        src = file.path("resources", "shared", "sweetalert-1.0.1",
-                        "js", "sweetalert.min.js")
-      ),
-      shiny::tags$link(
-        rel = "stylesheet",
-        href = file.path("resources", "shared", "sweetalert-1.0.1",
-                         "css", "sweetalert.min.css")
-      ),
-      shiny::tags$script(
-        src = file.path("resources", "shared", "swalservice",
-                        "swalservice.min.js")
-      ),
-      shiny::tags$script(
-        src = file.path("resources", "srcjs", "shinyalert.js")
-      ),
-      shiny::tags$link(
-        rel = "stylesheet",
-        href = file.path("resources", "css", "shinyalert.css")
-      )
+useShinyalert <- function(rmd = FALSE, force = FALSE) {
+  if (force) {
+    if (rmd) {
+      session <- getSession()
+      session$userData$.shinyalert_added <- TRUE
+    }
+    return(getDependencies(inline = rmd))
+  }
+  if (rmd) {
+    warning(
+      "Good news! You don't need to call `useShinyalert()` anymore. Please remove this line from your code.\n",
+      "If you really want to pre-load {shinyalert} to the UI for performance reasons, use:\n",
+      "\t`useShinyalert(rmd = TRUE, force = TRUE)`",
+      call. = FALSE
     )
-  )
+  } else {
+    warning(
+      "Good news! You don't need to call `useShinyalert()` anymore. Please remove this line from your code.\n",
+      "If you really want to pre-load {shinyalert} to the UI for performance reasons, use:\n",
+      "\t`useShinyalert(force = TRUE)`",
+      call. = FALSE,
+      immediate. = TRUE
+    )
+  }
+  invisible(NULL)
 }