+++ R/utils.R
@@ -7,3 +7,35 @@ getSession <- function() {
 
   session
 }
+
+getDependencies <- function(inline = FALSE) {
+  shiny::addResourcePath("shinyalert-assets",
+                         system.file("www", package = "shinyalert"))
+
+  insert_into_doc <- if (inline) shiny::tagList else shiny::tags$head
+
+  shiny::singleton(
+    insert_into_doc(
+      shiny::tags$script(
+        src = file.path("shinyalert-assets", "shared", "sweetalert-1.0.1",
+                        "js", "sweetalert.min.js")
+      ),
+      shiny::tags$link(
+        rel = "stylesheet",
+        href = file.path("shinyalert-assets", "shared", "sweetalert-1.0.1",
+                         "css", "sweetalert.min.css")
+      ),
+      shiny::tags$script(
+        src = file.path("shinyalert-assets", "shared", "swalservice",
+                        "swalservice.min.js")
+      ),
+      shiny::tags$script(
+        src = file.path("shinyalert-assets", "srcjs", "shinyalert.js")
+      ),
+      shiny::tags$link(
+        rel = "stylesheet",
+        href = file.path("shinyalert-assets", "css", "shinyalert.css")
+      )
+    )
+  )
+}