+++ NEWS.md
@@ -1,5 +1,6 @@
 # Unreleased version
 
+- **MAJOR CHANGE** It's no longer required to call `useShinyalert()` in the UI. (#47) 
 - Fixed bug where tall modals did not have a scrollbar and you couldn't see the entire content (#44)
 
 # shinyalert 2.0.0 (2020-09-11)